Add exit confirmation message
