+++ src/main/java/tutu/exception/InvalidInputException.java
@@ -11,17 +11,4 @@ public class InvalidInputException extends Exception {
     public InvalidInputException(String message) {
         super(message);
     }
-
-    /**
-     * Shows an alert message explaining error.
-     * @param e Exception thrown due to invalid input.
-     */
-    public static void showErrorAlert(Exception e) {
-        Alert alert = new Alert(Alert.AlertType.ERROR);
-        alert.setTitle("Error!");
-        alert.setHeaderText("Your input is invalid!");
-        alert.setContentText(e.getMessage());
-
-        alert.showAndWait();
-    }
 }
